requirements.txt
discord.py
Pillow
requests
ffmpeg-python